Abstract container modifiers to use a container-base modifier
This commit extracts some of the shared logic for container based
modifiers into a common container-base base_modifier. This allows more
explicit standardization between the container modifiers.
