+++ var/ramble/repos/builtin/modifiers/pyxis-enroot/modifier.py
@@ -11,11 +11,12 @@ import os
 from ramble.modkit import *
 from ramble.util.hashing import hash_file, hash_string
 from spack.util.path import canonicalize_path
+from ramble.base_mod.builtin.container_base import ContainerBase
 
 import llnl.util.filesystem as fs
 
 
-class PyxisEnroot(BasicModifier):
+class PyxisEnroot(ContainerBase):
     """Modifier to aid configuring pyxis-enroot based execution environments
 
     Pyxis is a container plugin for slurm developed by NVIDIA.
@@ -41,6 +42,8 @@ class PyxisEnroot(BasicModifier):
     - container_env_vars
     """
 
+    _runtime = "enroot"
+    _unsquash = "unsquashfs"
     container_extension = "sqsh"
 
     container_hash_file_extension = "sha256"
@@ -51,28 +54,15 @@ class PyxisEnroot(BasicModifier):
 
     maintainers("douglasjacobsen")
 
-    mode("standard", description="Standard execution mode for pyxis-enroot")
     mode(
         "no_provenance",
         description="Standard execution mode without provenance tracking",
     )
-    default_mode("standard")
 
-    required_variable("container_name")
-    required_variable("container_uri")
-
-    modifier_variable(
-        "container_mounts",
-        default="",
-        description="Comma delimited list of mount points for the container. Filled in by modifier",
-        modes=["standard", "no_provenance"],
-    )
-
-    modifier_variable(
-        "container_env_vars",
-        default="",
-        description="Comma delimited list of environments to import into container. Filled in by modifier",
-        modes=["standard", "no_provenance"],
+    required_variable(
+        "container_name",
+        description="The variable controls the name of the resulting container file. "
+        "It will be of the format {container_name}.{container_extension}.",
     )
 
     modifier_variable(
@@ -83,120 +73,11 @@ class PyxisEnroot(BasicModifier):
     )
 
     modifier_variable(
-        "container_extract_dir",
-        default="{workload_input_dir}",
-        description="Directory where the extracted paths will be stored",
-        modes=["standard", "no_provenance"],
-    )
-
-    modifier_variable(
         "container_path",
         default="{container_dir}/{container_name}." + container_extension,
         description="Full path to the container sqsh file",
         modes=["standard", "no_provenance"],
     )
-
-    modifier_variable(
-        "container_extract_paths",
-        default="[]",
-        description="List of paths to extract from the sqsh file into the {workload_input_dir}. "
-        + "Will have paths of {workload_input_dir}/enroot_extractions/{path_basename}",
-        modes=["standard", "no_provenance"],
-        track_used=False,
-    )
-
-    def __init__(self, file_path):
-        super().__init__(file_path)
-
-        self.enroot_runner = None
-        self.unsquashfs_runner = None
-
-    def _build_commands(self, dry_run=False):
-        """Construct command runners for enroot and unsquashfs"""
-        if self.enroot_runner is None:
-            self.enroot_runner = CommandRunner(
-                name="enroot", command="enroot", dry_run=dry_run
-            )
-
-        if self.unsquashfs_runner is None:
-            self.unsquashfs_runner = CommandRunner(
-                name="unsquashfs", command="unsquashfs", dry_run=dry_run
-            )
-
-    register_phase(
-        "define_container_variables",
-        pipeline="setup",
-        run_before=["get_inputs"],
-    )
-
-    def _define_container_variables(self, workspace, app_inst=None):
-        """Define helper variables for working with enroot experiments
-
-        To ensure it is defined properly, construct a comma delimited list of
-        environment variable names that will be added into the
-        container_env_vars variable.
-        """
-
-        def extract_names(itr, name_set=set()):
-            """Extract names of environment variables from the environment variable action sets
-
-            Given an iterator over environment variable action sets, extract
-            the names of the environment variables.
-
-            Modifies the name_set argument inplace.
-            """
-            for action, conf in itr:
-                if action in ["set", "unset"]:
-                    for name in conf:
-                        name_set.add(name)
-                elif action == "prepend":
-                    for group in conf:
-                        for name in group["paths"]:
-                            name_set.add(name)
-                elif action == "append":
-                    for group in conf:
-                        for name in group["vars"]:
-                            name_set.add(name)
-
-        # Only define variables if mode is standard
-        if self._usage_mode == "standard":
-            # Define container_env-vars
-            set_names = set()
-
-            for env_var_set in app_inst._env_variable_sets:
-                extract_names(env_var_set.items(), set_names)
-
-            for mod_inst in app_inst._modifier_instances:
-                extract_names(mod_inst.all_env_var_modifications(), set_names)
-
-            env_var_list = ",".join(set_names)
-            app_inst.define_variable("container_env_vars", env_var_list)
-
-            # Define container_mounts
-            input_mounts = app_inst.expander.expand_var("{container_mounts}")
-
-            exp_mount = "{experiment_run_dir}:{experiment_run_dir}"
-            expanded_exp_mount = app_inst.expander.expand_var(exp_mount)
-
-            if (
-                exp_mount not in input_mounts
-                and expanded_exp_mount not in input_mounts
-            ):
-                add_mod = self._usage_mode not in self.variable_modifications
-                add_mod = (
-                    add_mod
-                    or self._usage_mode in self.variable_modifications
-                    and "container_mounts"
-                    not in self.variable_modifications[self._usage_mode]
-                )
-                if add_mod:
-                    self.variable_modification(
-                        "container_mounts",
-                        modification=exp_mount,
-                        separator=",",
-                        method="append",
-                        mode=self._usage_mode,
-                    )
 
     register_phase(
         "import_sqsh",
@@ -212,7 +93,9 @@ class PyxisEnroot(BasicModifier):
         (using enroot) into the target container_dir.
         """
 
-        self._build_commands(workspace.dry_run)
+        self._build_runner(
+            runtime=self._runtime, app_inst=app_inst, dry_run=workspace.dry_run
+        )
 
         uri = self.expander.expand_var_name("container_uri")
 
@@ -242,7 +125,11 @@ class PyxisEnroot(BasicModifier):
     def _extract_from_sqsh(self, workspace, app_inst=None):
         """Extract paths from the sqsh file into the workload inputs path"""
 
-        self._build_commands(workspace.dry_run)
+        self._build_runner(
+            runtime=self._unsquash,
+            app_inst=app_inst,
+            dry_run=workspace.dry_run,
+        )
 
         extract_paths = self.expander.expand_var_name(
             "container_extract_paths", typed=True, merge_used_stage=False
